refactor(Alarms): changed name from 'AlarmInfoChart' to 'ChartAlarmsByMonth'
- Created hooks, utils and components folders.
- Moved responsibilities to the different folders and files.
