+++ strateger-react/src/components/Alarms/containers/AlarmInfoPanel.jsx
@@ -1,11 +1,11 @@
 //Path: src/components/Alarms/containers/AlarmInfoPanel.jsx
 
-import AlarmInfoChart from '../components/AlarmInfoPanel/AlarmInfoChart';
+import AlarmInfoChart from '../components/AlarmInfoPanel/ChartAlarmsByMonth/ChartAlarmsByMonth';
 
 const AlarmInfoPanel = () => {
     return(
-        <div className='h-full'>
-            <AlarmInfoChart />        
+        <div className='h-full bg-african_violet-200 p-2'>
+            <AlarmInfoChart />                    
         </div>
     )
 }
